moved test_arg_load to integration tests
